Split event router unit tests into separate files
